Remove support for multiple authors/affiliations/logos
It makes things complicated for the theme while not doing much more
that a user can do easily by giving a laid out box to `author` or
`affiliation`. And supporting properly the cases of multiple authors
with same and with different affiliation would be even more complicated.
And it's rarely used.